Add scripts to enable automatic layout generation
So far, we've got this up and running for Xfce.
MATE's on the roadmap, but probably a bit more intricate to get right,
so unlikely to come immediately.

I'm also contemplating a default configurator based on xrandr/xprop, but
not sure whether that would be a redundant implementation.

May end up making an openbox user very happy, idk.